+++ src/Nix/Normal.hs
@@ -140,6 +140,9 @@ normalForm_
   -> m ()
 normalForm_ t = void $ normalizeValue t
 
+opaqueVal :: Applicative f => NValue t f m
+opaqueVal = nvStr $ makeNixStringWithoutContext "<cycle>"
+
 -- | Detect cycles & stub them.
 stubCycles
   :: forall t f m
@@ -161,7 +164,10 @@ stubCycles =
     )
     Free
  where
-  Free (NValue' cyc) = opaque
+  Free (NValue' cyc) = opaqueVal
+
+thunkVal :: Applicative f => NValue t f m
+thunkVal = nvStr $ makeNixStringWithoutContext "<thunk>"
 
 removeEffects
   :: (MonadThunk t m (NValue t f m), MonadDataContext f m)
@@ -171,14 +177,11 @@ removeEffects =
   iterNValueM
     id
     --  2021-02-25: NOTE: Please, unflip this up the stack
-    (\ t f -> f =<< queryM (pure opaque) t)
+    (\ t f -> f =<< queryM (pure thunkVal) t)
     (fmap Free . sequenceNValue' id)
-
-opaque :: Applicative f => NValue t f m
-opaque = nvStr $ makeNixStringWithoutContext "<cycle>"
 
 dethunk
   :: (MonadThunk t m (NValue t f m), MonadDataContext f m)
   => t
   -> m (NValue t f m)
-dethunk = removeEffects <=< queryM (pure opaque)
+dethunk = removeEffects <=< queryM (pure thunkVal)